refactor(streamlit): extract landing and map sidebar + working set
Refs #131. Implemented in #126.

Made-with: Cursor
